refactor: use js extension instead mjs in tools and examples
