+++ .mocharc.cjs
@@ -3,5 +3,5 @@ module.exports = {
   recursive: true,
   extension: '.ts',
   timeout: process.env.NETWORK ? '30s' : '6s',
-  ignore: 'test/environment/**',
+  ignore: ['test/charts/**', 'test/emitter/**', 'test/environment/**'],
 };
