+++ examples/README.md
@@ -16,9 +16,9 @@ Create a [new issue](https://github.com/aeternity/aepp-sdk-js/issues/new) to sug
 
 ## NodeJS
 
-1. [Contract interaction](node/contract-interaction.mjs)
-2. [Transfer AE](node/transfer-ae.mjs)
-3. [Paying for spend tx](node/paying-for-spend-tx.mjs)
-4. [Paying for contract call tx](node/paying-for-contract-call-tx.mjs)
-5. [Dry-run using debug endpoint](node/dry-run-using-debug-endpoint.mjs)
-6. [Oracle](node/oracle.mjs)
+1. [Contract interaction](node/contract-interaction.js)
+2. [Transfer AE](node/transfer-ae.js)
+3. [Paying for spend tx](node/paying-for-spend-tx.js)
+4. [Paying for contract call tx](node/paying-for-contract-call-tx.js)
+5. [Dry-run using debug endpoint](node/dry-run-using-debug-endpoint.js)
+6. [Oracle](node/oracle.js)
