+++ examples/node/oracle.js
