+++ package.json
@@ -22,25 +22,25 @@
   },
   "sideEffects": false,
   "scripts": {
-    "build:assets": "node tooling/fetch-aesophia-cli.mjs",
+    "build:assets": "node tooling/fetch-aesophia-cli.js",
     "build:types": "tsc",
     "build:dist": "webpack",
     "build:es": "babel src --config-file ./babel.esm.config.js --out-dir es --extensions .ts --source-maps true",
-    "build:api:node": "autorest tooling/autorest/node.yaml && node tooling/autorest/postprocessing.mjs node",
-    "build:api:compiler": "node tooling/autorest/compiler-prepare.mjs && autorest tooling/autorest/compiler.yaml && node tooling/autorest/postprocessing.mjs compiler",
-    "build:api:middleware": "node tooling/autorest/middleware-prepare.mjs && autorest tooling/autorest/middleware.yaml && node tooling/autorest/postprocessing.mjs middleware",
+    "build:api:node": "autorest tooling/autorest/node.yaml && node tooling/autorest/postprocessing.js node",
+    "build:api:compiler": "node tooling/autorest/compiler-prepare.js && autorest tooling/autorest/compiler.yaml && node tooling/autorest/postprocessing.js compiler",
+    "build:api:middleware": "node tooling/autorest/middleware-prepare.js && autorest tooling/autorest/middleware.yaml && node tooling/autorest/postprocessing.js middleware",
     "build:generate": "tsx tooling/generate-schema.ts",
     "build": "run-p build:api:* build:assets build:generate && run-p build:dist build:es build:types",
-    "docs:examples": "node tooling/docs/examples-to-md.js examples/node/*.mjs",
+    "docs:examples": "node tooling/docs/examples-to-md.js examples/node/*.js",
     "docs:api": "typedoc",
     "commitlint": "commitlint --from develop",
     "lint": "run-p lint:*",
     "lint:prettier": "prettier . --check",
-    "lint:eslint": "eslint . --ext .ts,.js,.mjs",
+    "lint:eslint": "eslint . --ext .ts,.js",
     "lint:types": "tsc -p tsconfig.tests.json",
     "format": "prettier . --write",
     "test": "mocha './test/unit/' './test/integration/'",
-    "test:assets": "node tooling/fetch-metamask.mjs",
+    "test:assets": "node tooling/fetch-metamask.js",
     "test:integration": "mocha './test/integration/'",
     "test:unit": "mocha './test/unit/'",
     "test:watch": "mocha './test/unit/' './test/integration/' --watch",
