+++ test/emitter/Dockerfile
@@ -1,6 +1,6 @@
 FROM node:20-alpine
 WORKDIR /app
 
-COPY ./main.mjs main.mjs
+COPY ./main.js main.mjs
 
 CMD [ "node", "main.mjs" ]
